fix: Coreccion de errores  de las llaves foraneas y tambien en los endpoints

+++ src/main/java/com/pablosipac/kinalapp/repository/ClienteRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ClienteRepository extends JpaRepository<Cliente,String> {
-    List<Cliente> findByEstado (int Estado);
+    List<Cliente> findByEstado (Long Estado);
 }
